feat(wip): start adding plus button
